Break down big test file
